Update to current master of ATLauncher/ATLauncher

+++ src/main/java/com/atlauncher/gui/dialogs/EditModsDialog.java
@@ -24,18 +24,10 @@ import java.awt.event.ItemEvent;
 import java.awt.event.WindowAdapter;
 import java.awt.event.WindowEvent;
 import java.io.File;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Comparator;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.Optional;
 import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import javax.swing.AbstractButton;
 import javax.swing.JButton;
@@ -49,28 +41,20 @@ import javax.swing.JSplitPane;
 import javax.swing.SwingConstants;
 import javax.swing.UIManager;
 
+import org.mini2Dx.gettext.GetText;
+
 import com.atlauncher.App;
 import com.atlauncher.builders.HTMLBuilder;
 import com.atlauncher.data.DisableableMod;
 import com.atlauncher.data.Instance;
-import com.atlauncher.data.curseforge.CurseForgeFingerprint;
-import com.atlauncher.data.curseforge.CurseForgeProject;
-import com.atlauncher.data.minecraft.FabricMod;
-import com.atlauncher.data.minecraft.MCMod;
 import com.atlauncher.gui.components.ModsJCheckBox;
 import com.atlauncher.gui.handlers.ModsJCheckBoxTransferHandler;
 import com.atlauncher.gui.layouts.WrapLayout;
 import com.atlauncher.managers.ConfigManager;
 import com.atlauncher.managers.DialogManager;
-import com.atlauncher.managers.LogManager;
-import com.atlauncher.managers.PerformanceManager;
 import com.atlauncher.network.Analytics;
-import com.atlauncher.utils.CurseForgeApi;
 import com.atlauncher.utils.FileUtils;
-import com.atlauncher.utils.Hashing;
 import com.atlauncher.utils.Utils;
-
-import org.mini2Dx.gettext.GetText;
 
 public class EditModsDialog extends JDialog {
     private static final long serialVersionUID = 7004414192679481818L;
@@ -106,7 +90,7 @@ public class EditModsDialog extends JDialog {
 
         setupComponents();
 
-        scanMissingMods();
+        instance.scanMissingMods();
 
         loadMods();
 
@@ -246,7 +230,8 @@ public class EditModsDialog extends JDialog {
                             type = com.atlauncher.data.Type.shaderpack;
                         }
                         if (type != null) {
-                            DisableableMod mod = generateMod(file, type, App.settings.enableAddedModsByDefault);
+                            DisableableMod mod = DisableableMod.generateMod(file, type,
+                                    App.settings.enableAddedModsByDefault);
                             File copyTo = App.settings.enableAddedModsByDefault ? mod.getFile(instance)
                                     : mod.getDisabledFile(instance);
 
@@ -317,159 +302,9 @@ public class EditModsDialog extends JDialog {
         bottomPanel.add(closeButton);
     }
 
-    private DisableableMod generateMod(File file, com.atlauncher.data.Type type, boolean enabled) {
-        DisableableMod mod = new DisableableMod();
-        mod.disabled = !enabled;
-        mod.userAdded = true;
-        mod.wasSelected = true;
-        mod.file = file.getName();
-        mod.type = type;
-        mod.optional = true;
-        mod.name = file.getName();
-        mod.version = "Unknown";
-        mod.description = null;
-
-        MCMod mcMod = Utils.getMCModForFile(file);
-        if (mcMod != null) {
-            mod.name = Optional.ofNullable(mcMod.name).orElse(file.getName());
-            mod.version = Optional.ofNullable(mcMod.version).orElse("Unknown");
-            mod.description = Optional.ofNullable(mcMod.description).orElse(null);
-        } else {
-            FabricMod fabricMod = Utils.getFabricModForFile(file);
-            if (fabricMod != null) {
-                mod.name = Optional.ofNullable(fabricMod.name).orElse(file.getName());
-                mod.version = Optional.ofNullable(fabricMod.version).orElse("Unknown");
-                mod.description = Optional.ofNullable(fabricMod.description).orElse(null);
-            }
-        }
-
-        return mod;
-    }
-
-    private void scanMissingMods() {
-        PerformanceManager.start("EditModsDialog::scanMissingMods - CheckForAddedMods");
-
-        // files to scan
-        List<Path> files = new ArrayList<>();
-
-        // find the mods that have been added by the user manually
-        for (Path path : Arrays.asList(instance.ROOT.resolve("mods"), instance.ROOT.resolve("disabledmods"))) {
-            try (Stream<Path> stream = Files.list(path)) {
-                files.addAll(stream
-                        .filter(file -> !Files.isDirectory(file) && Utils.isAcceptedModFile(file)).filter(
-                                file -> instance.launcher.mods.stream()
-                                        .noneMatch(mod -> mod.type == com.atlauncher.data.Type.mods
-                                                && mod.file.equals(file.getFileName().toString())))
-                        .collect(Collectors.toList()));
-            } catch (IOException e) {
-                LogManager.logStackTrace("Error scanning missing mods", e);
-            }
-        }
-
-        if (files.size() != 0) {
-            final ProgressDialog progressDialog = new ProgressDialog(GetText.tr("Scanning New Mods"), 0,
-                    GetText.tr("Scanning New Mods"), this);
-
-            progressDialog.addThread(new Thread(() -> {
-                List<DisableableMod> mods = files.parallelStream()
-                        .map(file -> generateMod(file.toFile(), com.atlauncher.data.Type.mods,
-                                file.getParent().equals(instance.ROOT.resolve("mods"))))
-                        .collect(Collectors.toList());
-
-                if (!App.settings.dontCheckModsOnCurseForge) {
-                    Map<Long, DisableableMod> murmurHashes = new HashMap<>();
-
-                    mods.stream()
-                            .filter(dm -> dm.curseForgeProject == null && dm.curseForgeFile == null)
-                            .filter(dm -> dm.getFile(instance.ROOT, instance.id) != null).forEach(dm -> {
-                                try {
-                                    long hash = Hashing
-                                            .murmur(dm.disabled ? dm.getDisabledFile(instance).toPath()
-                                                    : dm
-                                                            .getFile(instance.ROOT, instance.id).toPath());
-                                    murmurHashes.put(hash, dm);
-                                } catch (Throwable t) {
-                                    LogManager.logStackTrace(t);
-                                }
-                            });
-
-                    if (murmurHashes.size() != 0) {
-                        CurseForgeFingerprint fingerprintResponse = CurseForgeApi
-                                .checkFingerprints(murmurHashes.keySet().stream().toArray(Long[]::new));
-
-                        if (fingerprintResponse != null && fingerprintResponse.exactMatches != null) {
-                            int[] projectIdsFound = fingerprintResponse.exactMatches.stream().mapToInt(em -> em.id)
-                                    .toArray();
-
-                            if (projectIdsFound.length != 0) {
-                                Map<Integer, CurseForgeProject> foundProjects = CurseForgeApi
-                                        .getProjectsAsMap(projectIdsFound);
-
-                                if (foundProjects != null) {
-                                    fingerprintResponse.exactMatches.stream()
-                                            .filter(em -> em != null && em.file != null
-                                                    && murmurHashes.containsKey(em.file.packageFingerprint))
-                                            .forEach(foundMod -> {
-                                                DisableableMod dm = murmurHashes
-                                                        .get(foundMod.file.packageFingerprint);
-
-                                                // add CurseForge information
-                                                dm.curseForgeProjectId = foundMod.id;
-                                                dm.curseForgeFile = foundMod.file;
-                                                dm.curseForgeFileId = foundMod.file.id;
-
-                                                CurseForgeProject curseForgeProject = foundProjects
-                                                        .get(foundMod.id);
-
-                                                if (curseForgeProject != null) {
-                                                    dm.curseForgeProject = curseForgeProject;
-                                                    dm.name = curseForgeProject.name;
-                                                    dm.description = curseForgeProject.summary;
-                                                }
-
-                                                LogManager.debug("Found matching mod from CurseForge called "
-                                                        + dm.curseForgeFile.displayName);
-                                            });
-                                }
-                            }
-                        }
-                    }
-                }
-
-                mods.forEach(mod -> LogManager.info("Found extra mod with name of " + mod.file));
-                instance.launcher.mods.addAll(mods);
-                instance.save();
-                progressDialog.close();
-            }));
-
-            progressDialog.start();
-        }
-        PerformanceManager.end("EditModsDialog::scanMissingMods - CheckForAddedMods");
-
-        PerformanceManager.start("EditModsDialog::scanMissingMods - CheckForRemovedMods");
-        // next remove any mods that the no longer exist in the filesystem
-        List<DisableableMod> removedMods = instance.launcher.mods.parallelStream().filter(mod -> {
-            if (!mod.wasSelected || mod.type != com.atlauncher.data.Type.mods) {
-                return false;
-            }
-
-            if (mod.disabled) {
-                return (mod.getFile(instance) != null && !mod.getDisabledFile(instance).exists());
-            } else {
-                return (mod.getFile(instance) != null && !mod.getFile(instance).exists());
-            }
-        }).collect(Collectors.toList());
-
-        if (removedMods.size() != 0) {
-            removedMods.forEach(mod -> LogManager.info("Mod no longer in filesystem: " + mod.file));
-            instance.launcher.mods.removeAll(removedMods);
-            instance.save();
-        }
-        PerformanceManager.end("EditModsDialog::scanMissingMods - CheckForRemovedMods");
-    }
-
     private void loadMods() {
         List<DisableableMod> mods = instance.launcher.mods.stream().filter(DisableableMod::wasSelected)
+                .filter(m -> !m.skipped)
                 .sorted(Comparator.comparing(m -> m.name)).collect(Collectors.toList());
         enabledMods = new ArrayList<>();
         disabledMods = new ArrayList<>();
@@ -545,7 +380,7 @@ public class EditModsDialog extends JDialog {
         progressDialog.addThread(new Thread(() -> {
             for (ModsJCheckBox mod : mods) {
                 if (mod.isSelected() && mod.getDisableableMod().isUpdatable()) {
-                    mod.getDisableableMod().checkForUpdate(this, instance);
+                    mod.getDisableableMod().checkForUpdate(progressDialog, instance);
                 }
                 progressDialog.doneTask();
             }
@@ -594,31 +429,40 @@ public class EditModsDialog extends JDialog {
     }
 
     private void removeMods() {
-        ArrayList<ModsJCheckBox> mods = new ArrayList<>(enabledMods);
-        for (ModsJCheckBox mod : mods) {
-            if (mod.isSelected()) {
-                this.instance.launcher.mods.remove(mod.getDisableableMod());
-                FileUtils.delete(
-                        (mod.getDisableableMod().isDisabled()
-                                ? mod.getDisableableMod().getDisabledFile(this.instance)
-                                : mod.getDisableableMod().getFile(this.instance)).toPath(),
-                        true);
-                enabledMods.remove(mod);
+        int ret = DialogManager.yesNoDialog(false)
+                .setTitle(GetText.tr("Delete Selected Mods?"))
+                .setContent(new HTMLBuilder().center().text(GetText.tr(
+                        "This will delete the selected mods from the instance.<br/><br/>Are you sure you want to do this?"))
+                        .build())
+                .setType(DialogManager.WARNING).show();
+
+        if (ret == 0) {
+            ArrayList<ModsJCheckBox> mods = new ArrayList<>(enabledMods);
+            for (ModsJCheckBox mod : mods) {
+                if (mod.isSelected()) {
+                    this.instance.launcher.mods.remove(mod.getDisableableMod());
+                    FileUtils.delete(
+                            (mod.getDisableableMod().isDisabled()
+                                    ? mod.getDisableableMod().getDisabledFile(this.instance)
+                                    : mod.getDisableableMod().getFile(this.instance)).toPath(),
+                            true);
+                    enabledMods.remove(mod);
+                }
             }
-        }
-        mods = new ArrayList<>(disabledMods);
-        for (ModsJCheckBox mod : mods) {
-            if (mod.isSelected()) {
-                this.instance.launcher.mods.remove(mod.getDisableableMod());
-                FileUtils.delete(
-                        (mod.getDisableableMod().isDisabled()
-                                ? mod.getDisableableMod().getDisabledFile(this.instance)
-                                : mod.getDisableableMod().getFile(this.instance)).toPath(),
-                        true);
-                disabledMods.remove(mod);
+            mods = new ArrayList<>(disabledMods);
+            for (ModsJCheckBox mod : mods) {
+                if (mod.isSelected()) {
+                    this.instance.launcher.mods.remove(mod.getDisableableMod());
+                    FileUtils.delete(
+                            (mod.getDisableableMod().isDisabled()
+                                    ? mod.getDisableableMod().getDisabledFile(this.instance)
+                                    : mod.getDisableableMod().getFile(this.instance)).toPath(),
+                            true);
+                    disabledMods.remove(mod);
+                }
             }
+            reloadPanels();
         }
-        reloadPanels();
     }
 
     public void reloadPanels() {
